+++ planetarium/gui/control_panel.gd
@@ -1,4 +1,4 @@
-# menu_panel.gd
+# control_panel.gd
 # This file is part of I, Voyager
 # https://ivoyager.dev
 # *****************************************************************************
@@ -17,24 +17,23 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # *****************************************************************************
-class_name MenuPanel
+class_name ControlPanel
 extends PanelContainer
+
+signal time_set_requested()
 
 
 func _ready():
-	var version_label = find_node("VersionLabel")
-	version_label.set_version_label("Planetarium", false, true)
-	var credits = find_node("Credits")
-	credits.set_hyperlink("Credits", "https://github.com/ivoyager/ivoyager/blob/master/CREDITS.md")
-	var feedback = find_node("Feedback")
-	feedback.set_hyperlink("Feedback", "https://www.ivoyager.dev/forum/")
-	var support_us = find_node("SupportUs")
-	support_us.set_hyperlink("Support Us!", "https://github.com/sponsors/ivoyager")
+	$"%SetDateTime".connect("pressed", self, "emit_signal", ["time_set_requested"])
+	$ControlDraggable.max_default_screen_proportions = Vector2(0.55, 0.45)
 	
+	# widget mods
+	$"%DateTimeLabel".clock_hms_format = "  %02d:%02d:%02d UT"
+	$"%DateTimeLabel".clock_hm_format = "  %02d:%02d UT"
 	$ControlDraggable.default_sizes = [
-		# Zeros allow panel to shrink to content, but we need some width here
-		# so our "Support Us!" RichTextLabel doesn't wrap.
-		Vector2(75.0, 0.0), # GUI_SMALL
-		Vector2(100.0, 0.0), # GUI_MEDIUM
-		Vector2(125.0, 0.0), # GUI_LARGE
+		Vector2(435.0, 0.0), # , 139.0), # GUI_SMALL
+		Vector2(575.0, 0.0), # , 168.0), # GUI_MEDIUM
+		Vector2(712.0, 0.0), # , 200.0), # GUI_LARGE
 	]
+
+