+++ planetarium/gui/gui_top.gd
@@ -2,7 +2,7 @@
 # This file is part of I, Voyager
 # https://ivoyager.dev
 # *****************************************************************************
-# Copyright 2017-2022 Charlie Whitfield
+# Copyright 2017-2023 Charlie Whitfield
 # I, Voyager is a registered trademark of Charlie Whitfield in the US
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -47,8 +47,8 @@ func _ready():
 		if !panel_container:
 			continue
 		panel_container.set("custom_styles/panel", style_box)
-	var set_date_time: Button = find_node("SetDateTime")
-	set_date_time.connect("pressed", $TimeSetPopup, "popup")
+	
+	$ControlPanel.connect("time_set_requested", $TimeSetPopup, "popup")
 
 
 func _on_project_builder_finished() -> void: