+++ planetarium/gui/info_panel.gd
@@ -2,7 +2,7 @@
 # This file is part of I, Voyager
 # https://ivoyager.dev
 # *****************************************************************************
-# Copyright 2017-2022 Charlie Whitfield
+# Copyright 2017-2023 Charlie Whitfield
 # I, Voyager is a registered trademark of Charlie Whitfield in the US
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -17,64 +17,81 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # *****************************************************************************
+class_name InfoPanel
 extends PanelContainer
 
 # Dynamically resizes to not grow bigger than data and not cover other panels.
 
 const MIN_DATA_SIZE := 80.0
 const UNDER_DATA_MARGIN := 20.0
+const UNDER_PANEL_GAP := 60.0
 
+var _world_targeting: Array = IVGlobal.world_targeting
 var _other_panels := [] # resize to not cover these
-var _suppress_resize := false
+var _suppress_resize := true
 
 onready var _selection_data: VBoxContainer = find_node("SelectionData")
 onready var _data_scroll: ScrollContainer = find_node("DataScroll")
 
 
 func _ready():
-	# widget mods
-	var date_time_label := find_node("DateTimeLabel")
-	date_time_label.clock_hms_format = "  %02d:%02d:%02d UT"
-	date_time_label.clock_hm_format = "  %02d:%02d UT"
-	var track_orbit_ground_ckbxs := find_node("TrackOrbitGroudCkbxs")
-	track_orbit_ground_ckbxs.remove_track_label()
+	IVGlobal.connect("simulator_started", self, "_on_simulator_started")
 	$ControlDraggable.default_sizes = [
-		Vector2(315.0, 0.0), #, 870.0), # GUI_SMALL
-		Vector2(375.0, 0.0), #, 1150.0), # GUI_MEDIUM
-		Vector2(455.0, 0.0), #, 1424.0), # GUI_LARGE
+		Vector2(315.0, 870.0), # GUI_SMALL
+		Vector2(375.0, 1150.0), # GUI_MEDIUM
+		Vector2(455.0, 1424.0), # GUI_LARGE
 	]
-	# y-axis resizing
-	_other_panels.append(get_parent().find_node("NavPanel"))
-	for other_panel in _other_panels:
-		other_panel.connect("item_rect_changed", self, "_resize")
-	connect("item_rect_changed", self, "_resize")
-	IVGlobal.connect("simulator_started", self, "_resize")
-	_selection_data.connect("resized", self, "_resize")
-	get_viewport().connect("size_changed", self, "_resize")
+	# limit panel bottom for other gui
+	connect("item_rect_changed", self, "_on_self_item_rect_changed")
+	for child in get_parent().get_children():
+		var control := child as Control
+		if !control or control == self:
+			continue
+		_other_panels.append(control)
+		control.connect("item_rect_changed", self, "_resize_vertical")
+	_selection_data.connect("resized", self, "_resize_vertical")
+	get_viewport().connect("size_changed", self, "_resize_vertical")
 
 
-func _resize() -> void:
+func _on_simulator_started() -> void:
+	_suppress_resize = false
+	_resize_vertical()
+
+
+func _on_self_item_rect_changed() -> void:
+	if _suppress_resize:
+		return
+	_resize_vertical()
+	_suppress_resize = true
+	yield(get_tree(), "idle_frame")
+	yield(get_tree(), "idle_frame")
+	_suppress_resize = false
+	_resize_vertical()
+
+
+func _resize_vertical() -> void:
 	if _suppress_resize:
 		return
 	_suppress_resize = true
 	var data_size := _selection_data.rect_size.y
 	var data_top := _data_scroll.get_global_rect().position.y
 	var above_data_size := data_top - rect_position.y
-	var bottom_limit := INF
 	var rect := get_rect()
-	for other_panel in _other_panels:
-		var other_rect: Rect2 = other_panel.get_rect()
+	var bottom_limit: float = _world_targeting[1] # Viewport height
+	for control in _other_panels:
+		if !control.visible:
+			continue
+		var other_rect: Rect2 = control.get_rect()
 		if rect.end.x < other_rect.position.x:
 			continue
 		if rect.position.x > other_rect.end.x:
 			continue
-		if bottom_limit > other_rect.position.y:
-			bottom_limit = other_rect.position.y
-	if bottom_limit == INF:
-		bottom_limit = get_viewport_rect().end.y
-	if data_size + data_top > bottom_limit: # grow to external limits, unless too small
-		var min_size := bottom_limit - rect_position.y
-		if min_size < above_data_size + MIN_DATA_SIZE:
+		var other_top: float = other_rect.position.y
+		if bottom_limit > other_top:
+			bottom_limit = other_top
+	if data_size + data_top > bottom_limit - UNDER_PANEL_GAP: # grow to external limits, unless too small
+		var min_size := bottom_limit - UNDER_PANEL_GAP - rect_position.y
+		if min_size < above_data_size + MIN_DATA_SIZE: # too small, just let them overlap
 			min_size = above_data_size + MIN_DATA_SIZE
 		rect_min_size.y = min_size
 	else: # shrink to content