+++ planetarium/gui/nav_panel.gd
@@ -2,7 +2,7 @@
 # This file is part of I, Voyager
 # https://ivoyager.dev
 # *****************************************************************************
-# Copyright 2017-2022 Charlie Whitfield
+# Copyright 2017-2023 Charlie Whitfield
 # I, Voyager is a registered trademark of Charlie Whitfield in the US
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -17,16 +17,18 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # *****************************************************************************
+class_name NavPanel
 extends PanelContainer
 
 
 var _settings: Dictionary = IVGlobal.settings
-var _under_moons_spacer_sizes := [55.0, 66.0, 77.0]
-
-onready var _under_moons_spacer: Control = find_node("UnderMoonsSpacer")
 
 
 func _ready():
+	# widgets
+	$"%AsteroidsHScroll".add_bodies_from_table("asteroids")
+	$"%SpacecraftHScroll".add_bodies_from_table("spacecrafts")
+	
 	$ControlDraggable.default_sizes = [
 		Vector2(435.0, 278.0), # GUI_SMALL
 		Vector2(575.0, 336.0), # GUI_MEDIUM
@@ -38,8 +40,8 @@ func _ready():
 
 
 func _resize() -> void:
-	var gui_size: int = _settings.gui_size
-	_under_moons_spacer.rect_min_size.y = _under_moons_spacer_sizes[gui_size]
+	pass
+#	var gui_size: int = _settings.gui_size
 
 
 func _settings_listener(setting: String, _value) -> void: